Repo cleanup: consolidate duplicates, remove dead code, update stale refs
Full-repo cleanup pass across backend, scripts, and tests.

SAFE REMOVALS:
- Remove duplicate utc_now_iso() from src/api/data_contract.py;
  import from src.data_models.contracts instead (single source of truth)
- Remove unused datetime/timezone import from data_contract.py

CONSOLIDATION:
- Create scripts/_shared.py with 3 helpers that were copy-pasted across
  5-10 scripts: _repo_root(), _latest(), _normalize_name()
- Update 10 scripts to import from _shared instead of defining locally
- Net: ~40 lines of duplicate code removed, single maintenance point
- Add scripts/__init__.py so the package is importable

TEST/DOC ALIGNMENT:
- Update test_calibration.py and test_enrich.py to reference current
  legacy data file (2026-03-22) instead of stale (2026-03-10)
- Update pick calibration assertion to match current data (6656 vs 7311)

VALIDATION:
- 459 backend tests pass
- 86 frontend tests pass
- All script imports verified working

INTENTIONALLY LEFT ALONE:
- Dynasty Scraper.py (11,787 lines) — legacy production backbone
- codex_loop.py, sync.bat — developer workflow tools
- trade-logic.js unused exports — tested utility library, likely
  planned for future refactor of trade/page.jsx
- Position normalization variants — context-specific, not true dupes
- data/_prev_composites.json, data/_latest_alert.txt — used by scraper
