+++ scripts/calibration_sweep.py
@@ -14,14 +14,12 @@ import json
 import sys
 from pathlib import Path
 
+# Ensure repo root is on sys.path for shared imports
+_REPO = Path(__file__).resolve().parents[1]
+if str(_REPO) not in sys.path:
+    sys.path.insert(0, str(_REPO))
 
-def _repo_root() -> Path:
-    return Path(__file__).resolve().parents[1]
-
-
-def _latest_file(directory: Path, pattern: str) -> Path | None:
-    files = sorted(directory.glob(pattern), reverse=True)
-    return files[0] if files else None
+from scripts._shared import _repo_root, _latest as _latest_file
 
 
 def recalibrate_assets(assets: list[dict], exponent: float, universe_scales: dict[str, int]) -> list[dict]:
